pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.302",
    "litellm>=1.65.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
]
